+++ backend/core/src/listings/types/listing-filter-keys-enum.ts
@@ -5,4 +5,5 @@ export enum ListingFilterKeys {
   neighborhood = "neighborhood",
   bedrooms = "bedrooms",
   zipcode = "zipcode",
+  seniorHousing = "seniorHousing",
 }