+++ backend/core/src/shared/filter/custom_filters.ts
@@ -1,0 +1,23 @@
+import { WhereExpression } from "typeorm"
+import { ListingFilterKeys } from "../../listings/types/listing-filter-keys-enum"
+import { filterTypeToFieldMap } from "../../listings/dto/listing.dto"
+
+export function addSeniorHousingQuery(qb: WhereExpression, filterValue: string) {
+  const whereParameterName = ListingFilterKeys.seniorHousing
+  const seniorHousingCommunityType = "senior62"
+  const reservedCommunityTypeColumnName = `LOWER(CAST(${
+    filterTypeToFieldMap[ListingFilterKeys.seniorHousing]
+  } as text))`
+  if (filterValue == "true") {
+    qb.andWhere(`${reservedCommunityTypeColumnName} = LOWER(:${whereParameterName})`, {
+      [whereParameterName]: seniorHousingCommunityType,
+    })
+  } else if (filterValue == "false") {
+    qb.andWhere(
+      `(${reservedCommunityTypeColumnName} IS NULL OR ${reservedCommunityTypeColumnName} <> LOWER(:${whereParameterName}))`,
+      {
+        [whereParameterName]: seniorHousingCommunityType,
+      }
+    )
+  }
+}