+++ backend/core/types/src/backend-swagger.ts
@@ -3714,6 +3714,9 @@ export interface ListingFilterParams {
 
   /**  */
   zipcode?: string
+
+  /** */
+  seniorHousing?: boolean
 }
 
 export interface UnitType {