+++ sites/public/pages/eligibility/income.tsx
@@ -10,17 +10,25 @@ import { ProgressNav } from "@bloom-housing/ui-components/src/navigation/Progres
 import { ELIGIBILITY_SECTIONS } from "../../lib/constants"
 import { Form } from "@bloom-housing/ui-components/src/forms/Form"
 import { Button } from "@bloom-housing/ui-components/src/actions/Button"
-import { AppearanceStyleType, Select } from "@bloom-housing/ui-components"
+import {
+  AppearanceStyleType,
+  encodeToFrontendFilterString,
+  Select,
+} from "@bloom-housing/ui-components"
 import { useForm } from "react-hook-form"
 import { EligibilityContext } from "../../lib/EligibilityContext"
 import { eligibilityRoute } from "../../lib/helpers"
 import FormBackLink from "../../src/forms/applications/FormBackLink"
+import { useRouter } from "next/router"
+import { ListingFilterParams } from "@bloom-housing/backend-core/types"
 
 const EligibilityIncome = () => {
+  const router = useRouter()
   const { eligibilityRequirements } = useContext(EligibilityContext)
 
   const incomeRanges = ["below10k", "10kTo20k", "30kTo40k", "40kTo50k", "over50k"]
   const CURRENT_PAGE = 4
+  const SENIOR_AGE = 62
 
   /* Form Handler */
   // eslint-disable-next-line @typescript-eslint/unbound-method
@@ -33,10 +41,21 @@ const EligibilityIncome = () => {
     const data = getValues()
     const { income } = data
     eligibilityRequirements.setIncome(income)
+    void router.push(getFilterUrl())
   }
 
   if (eligibilityRequirements.completedSections <= CURRENT_PAGE) {
     eligibilityRequirements.setCompletedSections(CURRENT_PAGE + 1)
+  }
+
+  function getFilterUrl() {
+    const params: ListingFilterParams = {}
+
+    if (eligibilityRequirements.age < SENIOR_AGE) {
+      params.seniorHousing = false
+    }
+
+    return `/listings?${encodeToFrontendFilterString(params)}`
   }
 
   return (